git-svn-id: https://faracvs.cs.uni-magdeburg.de/svn/koeppen-QuelltextProj/src@243 9da66865-35ea-40ca-919a-acafe8e375a9

+++ experimentGUI/util/experimentEditorDataVisualizer/Actions/Action.java
@@ -4,7 +4,7 @@ public interface Action extends Comparable<Integer>{
 	
 	public String getActionCommand();
 	
-	public int getStartTime();
+	public long getStartTime();
 	
 	public String getPath();
 }
